init
Initial commit, based on forking from DevOS[1] using Digga[2].

This configuration was initially created 2021/09/08

Over the course of a few from-root rebases history has been rewritten,
however we have tried to keep the changes consistent to represent the
work done.

Over time, we have removed commits we did not write in order to keep the
repository "purely" our work, but we shouldn't eliminate the shoulders of
the giant we stand on. This config was created with reference to:

- nrdxp, thank you for creating Digga and DevOS
- hlissner, who helped inspire Digga and whose dotfiles also inspired us
- numtime and contributors, for creating flake-utils and devshell
- gytis-invaskervicius, for creating flake-utils-plus
- All DevOS and Digga contributors, whose work helped us get started

While their names are no longer in commits in this repo, we would not be
where we are today without them.

~ Infinidoge, 2026/02/24

[1]: https://github.com/divnix/digga/tree/main/examples/devos
[2]: https://github.com/divnix/digga