fix(Infini-SERVER): disable apcupsd
Shutting down my desktop would likely cause apcupsd on the server to
freak out, and possibly shut down the computer on its own.
I'll likely reconfigure the setup to have Infini-SERVER be the primary
connection to the UPS, with the desktop acting as the secondary, as
Infini-SERVER is online all of the time where possible.